Moving random string generator into utils
